games/requirements.txt
numpy>=1.21.2
opencv-python>=4.5.3.56
pygame>=2.0.1
pyautogui==0.9.53
screeninfo==0.7
requests==2.26.0
mediapipe>=0.8.7.3
screeninfo>=0.7
cvzone==1.5.6
pyautogui==0.9.53
requests==2.26.0
pynput
selenium
pynput
websocket-client